Rename the browsercontext module.
